🚀 feat(server): error handling program arguments

+++ server/lib/get_size_word_array.c
@@ -9,8 +9,9 @@
 
 size_t get_size_word_array(char **array)
 {
-    int i = 0;
+    size_t i = 0;
 
     for (; array[i] != NULL; i++);
+
     return i;
 }
